Updated to tk 8.4.1

+++ tk/xlib/ximage.c
@@ -8,7 +8,7 @@
  * See the file "license.terms" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * RCS: @(#) $Id: ximage.c,v 1.7.6.1 2000/05/04 21:26:31 spolk Exp $
+ * RCS: @(#) $Id: ximage.c,v 1.4 2000/07/06 03:17:45 mo Exp $
  */
 
 #include "tkInt.h"
@@ -47,7 +47,7 @@ XCreateBitmapFromData(display, d, data, width, height)
     GC gc;
     Pixmap pix;
 
-    pix = Tk_GetPixmap(display, d, width, height, 1);
+    pix = Tk_GetPixmap(display, d, (int) width, (int) height, 1);
     gc = XCreateGC(display, pix, 0, NULL);
     if (gc == NULL) {
 	return None;
@@ -69,4 +69,3 @@ XCreateBitmapFromData(display, d, data, width, height)
     XFreeGC(display, gc);
     return pix;
 }
-
